requirements.txt
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary==2.9.5
pyparsing==3.0.9
pysolr==3.9.0
redis==4.3.4
requests==2.28.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
